feat: add support for breaking change in cypress
- No longer support calling `autoStub` from a lifecycle hook. Instead it must be caleld from the top-level of a describe block
